test/test Add function stubs for native and applicable (ignored) tests.
A significant refactor of the code was merged recently that implemented the cuda
back end. Native and OpenCL were left behind and commented out. Further,
building with `cargo test --no-default-features --feature native` failed because
cudann was still being pulled in by the tests. This commit:

1. fixes the cfg guards so dropping cuda support is possible for some people.

2. Adds the function stubs for the Native backend. These are left w/
notimplemented!();

3. The same tests for the cuda backend are copy/pasted and refactored to apply
to the Native backend.

This puts us in a good position for someone to just implement the Native backend
functions now.

NB: There is still a failing test: the lib.rs doc test requires Cuda. Adding a
cfg guard to it uglifies it too much for it to remain a useful example, so I
left it alone.