chore/cudnn: update api to use latest v5.1 cudnn
Note that this has still a few sore spots, namely
 * missing WINOGRAD implementation
 * missing WINOGRAD_FUSED implementation
 * init_activation() is called all over the place

There should be some consideration only exporting activation from cudnn
and just specializing either in collenchyma-nn or even passen through to the user.
The downside would be the fact that this would imply constraints on other
backends to implement the same thing. Not sure if that is what we want.